% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pikachur.R
\name{catchemall}
\alias{catchemall}
\title{Know your Pokemon Stats: Where're my Pokemon party monsters at?!}
\usage{
catchemall(pokemon1, pokemon2, pokemon3, pokemon4, pokemon5, pokemon6)
}
\arguments{
\item{pokemon1}{Name of Pokemon.}

\item{pokemon2}{Name of Pokemon.}

\item{pokemon3}{Name of Pokemon.}

\item{pokemon4}{Name of Pokemon.}

\item{pokemon5}{Name of Pokemon.}

\item{pokemon6}{Name of Pokemon.}
}
\value{
data frame
}
\description{
In the Poke-universe, every Pokemon trainer can only hold a total of 6 Pokemon while on the go, also known as a Pokemon party. This function allows the Pokemon trainer to
get to know his/her party Pokemon (i.e. basic stats). This is advantageous for Pokemon trainers to understand each of their Pokemon's strengths and weaknesses in terms of their HP (hit points), attack and defense profiles.
}
\details{
This function asks the Pokemon player for the names of 6 of the Pokemon in his/her party. The output is a visualized barplot of the Pokemons' basic stats.
}
\examples{
catchemall("bulbasaur", "togepi", "pikachu", "squirtle", "cyndaquil", "psyduck")
# add 6 names of Pokemon
}
\author{
Montsie Guerrero <\url{http://github.com/acg2176}>
}
