% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pikachur.R
\name{pokevolve}
\alias{pokevolve}
\title{Ready, Set, Evolve! Pokemon, All Grown Up"}
\usage{
pokevolve(pokemon1, pokemon2, pokemon3)
}
\arguments{
\item{pokemon1}{Name of Pokemon.}

\item{pokemon2}{Name of Pokemon.}

\item{pokemon3}{Name of Pokemon.}
}
\value{
data frame
}
\description{
Like a proud parent, all Pokemon trainers beam with pride when their Pokemon make it to the next evolution. Happy tears aside, this function allows the Pokemon trainer
to determine how their Pokemon evolve (i.e. by level, by special item, etc). The output of this function is a dataset that provides information on the Pokemon evolution chain.
}
\details{
The function takes in 3 arguments. Pokemon trainers can select 3 Pokemon to find out each Pokemon's evolution chain. If Pokemon that is inputted does not have an evolution chain, a warning message will occur.
}
\examples{
pokevolve("pikachu","weedle","blastoise")
# add 3 Pokemon names
}
\author{
Montsie Guerrero <\url{http://github.com/acg2176}>
}
